requirements.txt
fastapi
fastmcp
langchain
langgraph
langchain_core
langchain_openai
langchain_community
langsmith
python-dotenv
uvicorn
bcrypt
pydantic
requests
pyjwt
sentence-transformers
chromadb
pinecone
langchain-huggingface
langchain-text-splitters
motor
python-socketio
pymongo
sqlalchemy
asyncmy
sqlmodel
pymysql
fastapi-mail